requirements.txt
django==3.2
djangorestframework==3.11
djoser
django-cors-headers
celery[redis]
django-celery-beat
lxml
requests
